remove renderingmainwindowcustomizer
