--- apps/rendering/gui/controller/renderingmainwindowcustomizer.py
@@ -1,167 +1,0 @@
-import datetime
-import logging
-import os
-import time
-from PyQt4 import QtCore
-
-from PyQt4.QtGui import QPixmap, QTreeWidgetItem, QPainter, QColor, QPen, QMessageBox, QPixmapCache
-
-from golem.core.common import get_golem_path
-
-
-from apps.core.task.gnrtaskstate import TaskDesc
-from apps.rendering.gui.controller.renderingnewtaskdialogcustomizer import RenderingNewTaskDialogCustomizer
-from apps.rendering.task.framerenderingtask import get_frame_name
-from apps.rendering.task.renderingtaskstate import RenderingTaskDefinition
-
-from gui.controller.mainwindowcustomizer import MainWindowCustomizer
-from gui.controller.memoryhelper import resource_size_to_display, translate_resource_index
-from gui.controller.showtaskresourcesdialogcustomizer import ShowTaskResourcesDialogCustomizer
-from gui.guidirmanager import get_preview_file
-from gui.view.dialog import ShowTaskResourcesDialog
-
-
-logger = logging.getLogger("apps.rendering")
-
-frame_renderers = [u"Blender"]
-
-
-def insert_item(root, path_table):
-    assert isinstance(root, QTreeWidgetItem)
-
-    if len(path_table) > 0:
-        for i in range(root.childCount()):
-            if path_table[0] == "{}".format(root.child(i).text(0)):
-                insert_item(root.child(i), path_table[1:])
-                return
-
-        new_child = QTreeWidgetItem([path_table[0]])
-        root.addChild(new_child)
-        insert_item(new_child, path_table[1:])
-
-
-class AbsRenderingMainWindowCustomizer(object):
-
-
-    def _setup_advance_task_connections(self):
-        self.gui.ui.showResourceButton.clicked.connect(self._show_task_resource_clicked)
-
-    def _set_new_task_dialog_customizer(self):
-        self.new_task_dialog_customizer = RenderingNewTaskDialogCustomizer(self.gui, self.logic)
-
-    def _set_show_task_resource_dialog(self):
-        self.show_task_resources_dialog = ShowTaskResourcesDialog(self.gui.window)
-        show_task_resources_dialog_customizer = ShowTaskResourcesDialogCustomizer(self.show_task_resources_dialog, self)
-
-    def update_task_additional_info(self, t):
-        assert isinstance(t, TaskDesc)
-
-        self.current_task_highlighted = t
-        self.__set_time_params(t)
-
-        if not isinstance(t.definition, RenderingTaskDefinition):
-            return
-
-        self.__set_renderer_params(t)
-
-        if t.definition.task_type in frame_renderers and t.definition.options.use_frames:
-            self.__set_frame_preview(t)
-        else:
-            self.__set_preview(t)
-
-        self.__update_output_file_color()
-
-    def show_task_result(self, task_id):
-        t = self.logic.get_task(task_id)
-        if t.definition.task_type in frame_renderers and t.definition.options.use_frames:
-            file_ = self.__get_frame_name(t.definition, self.gui.ui.previewsSlider.value() - 1)
-        else:
-            file_ = t.definition.output_file
-        if os.path.isfile(file_):
-            self.show_file(file_)
-        else:
-            self.show_error_window(u"{} is not a file".format(file_))
-
-    def __set_time_params(self, t):
-        self.gui.ui.subtaskTimeout.setText("{} minutes".format(int(t.definition.subtask_timeout / 60.0)))
-        self.gui.ui.fullTaskTimeout.setText(str(datetime.timedelta(seconds=t.definition.full_task_timeout)))
-        if t.task_state.time_started != 0.0:
-            lt = time.localtime(t.task_state.time_started)
-            time_string = time.strftime("%Y.%m.%d  %H:%M:%S", lt)
-            self.gui.ui.timeStarted.setText(time_string)
-
-    def __set_renderer_params(self, t):
-        mem, index = resource_size_to_display(t.definition.estimated_memory / 1024)
-        self.gui.ui.estimatedMemoryLabel.setText("{} {}".format(mem, translate_resource_index(index)))
-
-    def __set_frame_preview(self, t):
-        if "resultPreview" in t.task_state.extra_data:
-            self.slider_previews = t.task_state.extra_data["resultPreview"]
-        self.gui.ui.previewsSlider.setVisible(True)
-        self.gui.ui.previewsSlider.setRange(1, len(t.definition.options.frames))
-        self.gui.ui.previewsSlider.setSingleStep(1)
-        self.gui.ui.previewsSlider.setPageStep(1)
-        self._update_slider_preview()
-        frame_num = self.__get_frame_name(t.definition, self.gui.ui.previewsSlider.value() - 1)
-        self.gui.ui.outputFile.setText(u"{}".format(frame_num))
-
-    def __set_preview(self, t):
-        self.gui.ui.outputFile.setText(u"{}".format(t.definition.output_file))
-        self.gui.ui.previewsSlider.setVisible(False)
-        if "resultPreview" in t.task_state.extra_data and os.path.exists(os.path.abspath(t.task_state.extra_data["resultPreview"])):
-            file_path = os.path.abspath(t.task_state.extra_data["resultPreview"])
-            self.preview_controller.update_img(QPixmap(file_path))
-            self.last_preview_path = file_path
-        else:
-            self.preview_path = get_preview_file()
-            self.preview_controller.update_img(QPixmap(self.preview_path))
-            self.last_preview_path = self.preview_path
-
-    @staticmethod
-    def __get_frame_name(definition, num):
-        output_name, ext = os.path.splitext(definition.output_file)
-        frame_num = definition.options.frames[num]
-        return get_frame_name(output_name, ext[1:], frame_num)
-
-    def __update_output_file_color(self):
-        if os.path.isfile(self.gui.ui.outputFile.text()):
-            self.gui.ui.outputFile.setStyleSheet('color: blue')
-        else:
-            self.gui.ui.outputFile.setStyleSheet('color: black')
-
-    def _show_task_resource_clicked(self):
-
-        if self.current_task_highlighted:
-            res = [os.path.abspath(r) for r in self.current_task_highlighted.definition.resources]
-            res.sort()
-            self._set_show_task_resource_dialog()
-
-            item = QTreeWidgetItem(["Resources"])
-            self.show_task_resources_dialog.ui.folderTreeView.insertTopLevelItem(0, item)
-            self.show_task_resources_dialog.ui.closeButton.clicked.connect(self.__show_task_res_close_button_clicked)
-
-            for r in res:
-                after_split = r.split("\\")
-                insert_item(item, after_split)
-
-            self.show_task_resources_dialog.ui.mainSceneFileLabel.setText(
-                self.current_task_highlighted.definition.main_scene_file)
-            self.show_task_resources_dialog.ui.folderTreeView.expandAll()
-
-            self.show_task_resources_dialog.show()
-
-    def __show_task_res_close_button_clicked(self):
-        self.show_task_resources_dialog.window.close()
-
-
-class RenderingMainWindowCustomizer(AbsRenderingMainWindowCustomizer, MainWindowCustomizer):
-    def __init__(self, gui, logic):
-        MainWindowCustomizer.__init__(self, gui, logic)
-        self._setup_advance_task_connections()
-
-    def init_config(self):
-        MainWindowCustomizer.init_config(self)
-
-    def update_task_additional_info(self, t):
-        MainWindowCustomizer.update_task_additional_info(self, t)
-        AbsRenderingMainWindowCustomizer.update_task_additional_info(self, t)
